+++ packages/node-core/src/configure/ProjectUpgrade.service.ts
@@ -3,16 +3,27 @@
 
 import assert from 'assert';
 import {isMainThread} from 'worker_threads';
-import {findLast, last} from 'lodash';
-import {CacheMetadataModel, ISubqueryProject} from '../indexer';
+import {getAllEntitiesRelations} from '@subql/utils';
+import {Sequelize, Transaction} from '@subql/x-sequelize';
+import {findLast, last, parseInt} from 'lodash';
+import {ISubqueryProject, StoreCacheService, StoreService} from '../indexer';
 import {getLogger} from '../logger';
 import {getStartHeight, mainThreadOnly} from '../utils';
 import {BlockHeightMap} from '../utils/blockHeightMap';
+import {SchemaMigrationService} from './migration-service';
+import {NodeConfig} from './NodeConfig';
 
 type OnProjectUpgradeCallback<P> = (height: number, project: P) => void | Promise<void>;
 
 export interface IProjectUpgradeService<P extends ISubqueryProject = ISubqueryProject> {
-  init: (metadata: CacheMetadataModel, onProjectUpgrade?: OnProjectUpgradeCallback<P>) => Promise<number | undefined>;
+  init: (
+    storeCacheService: StoreCacheService,
+    currentHeight: number,
+    config: NodeConfig,
+    sequelize: Sequelize,
+    schema: string,
+    onProjectUpgrade?: OnProjectUpgradeCallback<P>
+  ) => Promise<number | undefined>;
   /**
    * This should only be called from within a worker thread as they dont have access to the store
    * */
@@ -20,9 +31,16 @@ export interface IProjectUpgradeService<P extends ISubqueryProject = ISubqueryPr
   updateIndexedDeployments: (id: string, blockHeight: number) => Promise<void>;
   readonly currentHeight: number;
   setCurrentHeight: (newHeight: number) => Promise<void>;
+  isRewindable: boolean;
   currentProject: P;
   projects: Map<number, P>;
   getProject: (height: number) => P;
+  rewind: (
+    targetBlockHeight: number,
+    lastProcessedHeight: number,
+    transaction: Transaction,
+    storeService: StoreService
+  ) => Promise<void>;
 }
 
 const serviceKeys: Array<keyof IProjectUpgradeService> = [
@@ -32,6 +50,7 @@ const serviceKeys: Array<keyof IProjectUpgradeService> = [
   'projects',
   'init',
   'updateIndexedDeployments',
+  'isRewindable',
 ];
 
 function assertEqual<T>(valueA: T, valueB: T, name: string) {
@@ -66,7 +85,7 @@ export function upgradableSubqueryProject<P extends ISubqueryProject>(
        * nestjs scheduler reflecting and being unable to get props.
        * If we can fix that then we can move this above the project accessors
        */
-      if (target[prop as keyof IProjectUpgradeService<P>] /*(serviceKeys as Array<string | symbol>).includes(prop)*/) {
+      if (prop in target /*(serviceKeys as Array<string | symbol>).includes(prop)*/) {
         const result = target[prop as keyof IProjectUpgradeService<P>];
 
         if (result instanceof Function) {
@@ -83,13 +102,14 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
   #currentHeight: number;
   #currentProject: P;
 
-  #metadata?: CacheMetadataModel;
-
+  #storeCache?: StoreCacheService;
   #initialized = false;
 
+  private config?: NodeConfig;
   private onProjectUpgrade?: OnProjectUpgradeCallback<P>;
+  private migrationService?: SchemaMigrationService;
 
-  private constructor(private _projects: BlockHeightMap<P>, currentHeight: number) {
+  private constructor(private _projects: BlockHeightMap<P>, currentHeight: number, private _isRewindable = true) {
     logger.info(
       `Projects: ${JSON.stringify(
         [..._projects.getAll().entries()].reduce((acc, curr) => {
@@ -100,14 +120,16 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
         2
       )}`
     );
-
     // Bypass setters here because we want to avoid side-effect
     this.#currentHeight = currentHeight;
     this.#currentProject = this.getProject(this.#currentHeight);
   }
-
   async init(
-    metadata: CacheMetadataModel,
+    storeCacheService: StoreCacheService,
+    currentHeight: number,
+    config: NodeConfig,
+    sequelize: Sequelize,
+    schema: string,
     onProjectUpgrade?: OnProjectUpgradeCallback<P>
   ): Promise<number | undefined> {
     if (this.#initialized) {
@@ -115,18 +137,25 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
       return;
     }
     this.#initialized = true;
-    this.#metadata = metadata;
-    this.onProjectUpgrade = onProjectUpgrade;
+    this.#storeCache = storeCacheService;
+    this.config = config;
+
+    this.migrationService = new SchemaMigrationService(
+      sequelize,
+      storeCacheService._flushCache.bind(storeCacheService),
+      schema,
+      config
+    );
 
     const indexedDeployments = await this.getDeploymentsMetadata();
 
     const lastProjectChange = this.validateIndexedData(indexedDeployments);
 
-    if (lastProjectChange) {
-      this.#currentHeight = lastProjectChange;
-      this.#currentProject = this.getProject(this.#currentHeight);
-    }
+    this.#currentHeight = lastProjectChange || currentHeight;
+    this.#currentProject = this.getProject(this.#currentHeight);
 
+    // executed last to ensure that the correct project is set first
+    this.onProjectUpgrade = onProjectUpgrade;
     return lastProjectChange;
   }
 
@@ -135,6 +164,10 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
     this.#currentHeight = currentHeight;
     this.#currentProject = this.getProject(this.#currentHeight);
     this.onProjectUpgrade = onProjectUpgrade;
+  }
+
+  get isRewindable(): boolean {
+    return this._isRewindable;
   }
 
   get projects(): Map<number, P> {
@@ -148,6 +181,61 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
   get currentHeight(): number {
     return this.#currentHeight;
   }
+  async rewind(
+    targetBlockHeight: number,
+    lastProcessedHeight: number,
+    transaction: Transaction,
+    storeService: StoreService
+  ): Promise<void> {
+    const projectsWithinRange = new BlockHeightMap(this.projects).getWithinRange(
+      targetBlockHeight,
+      lastProcessedHeight
+    );
+    const sortedProjectIds = Array.from(projectsWithinRange.keys());
+    const iterator = sortedProjectIds[Symbol.iterator]();
+
+    let currentId = iterator.next();
+    let nextId = iterator.next();
+
+    while (!nextId.done) {
+      const currentProject = projectsWithinRange.get(currentId.value);
+      const nextProject = projectsWithinRange.get(nextId.value);
+
+      if (currentProject && nextProject) {
+        if (this.config?.dbSchema) {
+          await storeService.init(getAllEntitiesRelations(currentProject.schema), this.config.dbSchema);
+        }
+        await this.migrate(currentProject, nextProject, transaction);
+      }
+
+      currentId = nextId;
+      this.#currentProject = nextProject as P;
+      nextId = iterator.next();
+    }
+  }
+
+  private async migrate(
+    project: ISubqueryProject,
+    newProject: ISubqueryProject,
+    transaction: Transaction | undefined
+  ): Promise<void> {
+    assert(this.config, 'NodeConfig is undefined');
+    if (!this.config.unfinalizedBlocks) {
+      assert(this.migrationService, 'MigrationService is undefined');
+      if (this.config.allowSchemaMigration) {
+        const modifiedModels = await this.migrationService.run(
+          project.schema,
+          newProject.schema,
+          this.#currentHeight,
+          transaction
+        );
+
+        if (modifiedModels) {
+          this.#storeCache?.updateModels(modifiedModels);
+        }
+      }
+    }
+  }
 
   async setCurrentHeight(height: number): Promise<void> {
     this.#currentHeight = height;
@@ -157,21 +245,24 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
 
     const hasChanged = this.#currentProject !== newProject;
     // Need to set this so that operations under hasChanged use the new project
+    const project = this.#currentProject;
     this.#currentProject = newProject;
 
     if (hasChanged) {
       if (isMainThread) {
         try {
-          // Use the project start height here for when resuming indexing at an arbitrary height
           await this.updateIndexedDeployments(newProject.id, startHeight);
         } catch (e: any) {
-          logger.error(e, 'Failed to update deployment metadata');
-          process.exit(1);
+          logger.error(e, 'Failed to update deployment');
+          throw e;
         }
       }
 
       try {
         await this.onProjectUpgrade?.(startHeight, newProject);
+        if (isMainThread) {
+          await this.migrate(project, newProject, undefined);
+        }
       } catch (e: any) {
         logger.error(e, `Failed to complete upgrading project`);
         process.exit(1);
@@ -190,6 +281,8 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
 
     let currentProject = startProject;
     let currentHeight: number | undefined;
+
+    let nextProject = startProject;
 
     const addProject = (height: number, project: P) => {
       this.validateProject(startProject, project);
@@ -223,7 +316,7 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
       }
 
       // Load the next project and repeat
-      const nextProject = await loadProject(currentProject.parent.reference).catch((e) => {
+      nextProject = await loadProject(currentProject.parent.reference).catch((e) => {
         throw new Error(`Failed to load parent project with cid: ${currentProject.parent?.reference}. ${e}`);
       });
       if (nextProject.parent && nextProject.parent.block > currentProject.parent.block) {
@@ -238,12 +331,27 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
       throw new Error('No valid projects found, this could be due to the startHeight.');
     }
 
+    const isRewindable = this.rewindableCheck(projects);
+
     assert(currentHeight, 'Unable to determine current height from projects');
-    return new ProjectUpgradeSevice(new BlockHeightMap(projects), currentHeight);
+    return new ProjectUpgradeSevice(new BlockHeightMap(projects), currentHeight, isRewindable);
   }
 
   getProject(height: number): P {
     return this._projects.get(height);
+  }
+
+  private static rewindableCheck<P extends ISubqueryProject>(projects: Map<number, P>): boolean {
+    const sortedProjects = new Map([...projects.entries()].sort((a, b) => a[0] - b[0]));
+    const projectIterator = sortedProjects.values();
+    let previousProject = projectIterator.next().value;
+    for (const project of projectIterator) {
+      if (!SchemaMigrationService.validateSchemaChanges(previousProject.schema, project.schema)) {
+        return false;
+      }
+      previousProject = project;
+    }
+    return true;
   }
 
   private static validateProject<P extends ISubqueryProject>(startProject: P, parentProject: P): void {
@@ -255,7 +363,7 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
 
   // Returns a height to rewind to if rewind is needed. Otherwise throws an error
   validateIndexedData(deploymentsMetadata: Record<number, string>): number | undefined {
-    // Using project upgades feature
+    // Using project upgrades feature
     if (this.projects.size > 1) {
       const indexedEntries = Object.entries(deploymentsMetadata);
       const projectEntries: [number, string][] = [...this.projects.entries()].map(([height, project]) => [
@@ -285,8 +393,8 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
   }
 
   private async getDeploymentsMetadata(): Promise<Record<number, string>> {
-    assert(this.#metadata, 'Project Upgrades service has not been initialized, unable to update metadata');
-    const deploymentsRaw = await this.#metadata.find('deployments');
+    assert(this.#storeCache?.metadata, 'Project Upgrades service has not been initialized, unable to update metadata');
+    const deploymentsRaw = await this.#storeCache?.metadata.find('deployments');
 
     if (!deploymentsRaw) return {};
 
@@ -295,7 +403,7 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
 
   @mainThreadOnly()
   async updateIndexedDeployments(id: string, blockHeight: number): Promise<void> {
-    assert(this.#metadata, 'Project Upgrades service has not been initialized, unable to update metadata');
+    assert(this.#storeCache?.metadata, 'Project Upgrades service has not been initialized, unable to update metadata');
     const deployments = await this.getDeploymentsMetadata();
 
     // If the last deployment is the same as the one we're updating to theres no need to do anything
@@ -313,6 +421,6 @@ export class ProjectUpgradeSevice<P extends ISubqueryProject = ISubqueryProject>
 
     deployments[blockHeight] = id;
 
-    this.#metadata.set('deployments', JSON.stringify(deployments));
+    this.#storeCache?.metadata.set('deployments', JSON.stringify(deployments));
   }
 }